+++ packages/flow-md-editor/src/index.ts
@@ -1,0 +1,1 @@
+export * from "./components/f-md-editor/f-md-editor";