+++ packages/flow-table/src/components/f-table-schema/f-table-schema.ts
@@ -6,14 +6,18 @@ import { ifDefined } from "lit-html/directives/if-defined.js";
 import { property, query, state } from "lit/decorators.js";
 import { FTable, FTableSelectable, FTableSize, FTableVariant } from "../f-table/f-table";
 import { FTcell, FTcellActions } from "../f-tcell/f-tcell";
-import { FTrow, FTrowState } from "../f-trow/f-trow";
+import { FTrow, FTrowChevronPosition, FTrowState } from "../f-trow/f-trow";
 import eleStyle from "./f-table-schema.scss";
+import { repeat } from "lit/directives/repeat.js";
 
 export type FTableSchemaDataRow = {
 	selected?: boolean;
 	details?: () => HTMLTemplateResult;
 	state?: FTrowState;
 	open?: boolean;
+	id: string;
+	disableSelection?: boolean;
+	expandIconPosition?: FTrowChevronPosition;
 	data: Record<string, FTableSchemaCell>;
 };
 export type FTableSchemaData = {
@@ -33,6 +37,7 @@ export type FTableSchemaHeaderCell = {
 	template?: () => HTMLTemplateResult;
 	width?: string;
 	selected?: boolean;
+	disableSort?: boolean;
 	sticky?: boolean;
 };
 
@@ -52,7 +57,8 @@ export class FTableSchema extends FRoot {
 		...FTrow.styles,
 		...FButton.styles,
 		...FSearch.styles,
-		...FText.styles
+		...FText.styles,
+		...FDiv.styles
 	];
 
 	/**
@@ -117,10 +123,19 @@ export class FTableSchema extends FRoot {
 	@property({ type: String, reflect: true, attribute: "search-term" })
 	searchTerm: string | null = null;
 	/**
+	 * search on selected header
+	 */
+	@property({ type: String, reflect: true, attribute: "search-scope" })
+	searchScope = "all";
+	/**
 	 * show search input box on top
 	 */
 	@property({ type: Boolean, reflect: true, attribute: "show-search-bar" })
 	showSearchBar = true;
+
+	set ["show-search-bar"](val: boolean) {
+		this.showSearchBar = val;
+	}
 
 	@state()
 	offset = 0;
@@ -150,15 +165,16 @@ export class FTableSchema extends FRoot {
 						}
 
 						return html`<f-tcell
+							part="cell"
 							.selected=${selected}
 							.width=${width}
 							?sticky-left=${ifDefined(sticky)}
 							?sticky-top=${ifDefined(this.stickyHeader)}
 							@selected-column=${this.handleColumnSelection}
 						>
-							<f-div gap="small" width="fit-content">
-								<f-text>${this.getHeaderCellTemplate(columnHeader[1])}</f-text>
-								${this.getSortIcon(columnHeader[0])}</f-div
+							<f-div gap="small" height="100%" width="fit-content">
+								${this.getHeaderCellTemplate(columnHeader[1])}
+								${columnHeader[1].disableSort ? nothing : this.getSortIcon(columnHeader[0])}</f-div
 							></f-tcell
 						>`;
 					})}
@@ -167,50 +183,64 @@ export class FTableSchema extends FRoot {
 	}
 
 	get rowsHtml() {
-		return this.filteredRows.map(row => {
-			const getDetailsSlot = () => {
-				if (row.details) {
-					return html` <f-div slot="details"> ${row.details()} </f-div>`;
-				} else {
-					return nothing;
-				}
-			};
-			return html`<f-trow
-				.open=${row.open ?? false}
-				.selected=${row.selected ?? false}
-				.state=${row.state ?? "default"}
-				@toggle-row=${(e: CustomEvent) => this.toggleRowDetails(row, e)}
-				@selected-row=${(e: CustomEvent) => this.handleRowSelection(row, e)}
-			>
-				${getDetailsSlot()}
-				${Object.entries(this.data.header).map(columnHeader => {
-					let width = undefined;
-					let selected = false;
-					let sticky = undefined;
-					let actions = undefined;
-					if (typeof columnHeader[1] === "object") {
-						if (columnHeader[1].width) {
-							width = columnHeader[1].width;
-						}
-						selected = columnHeader[1].selected ?? false;
-						sticky = columnHeader[1].sticky;
+		return repeat(
+			this.filteredRows,
+			row => row.id,
+			row => {
+				const getDetailsSlot = () => {
+					if (row.details) {
+						return html` <f-div slot="details" width="100%"> ${row.details()} </f-div>`;
+					} else {
+						return nothing;
 					}
-					const cell = row.data[columnHeader[0]];
+				};
+				return html`<f-trow
+					id=${row.id}
+					.expandIconPosition=${row.expandIconPosition ?? "right"}
+					.open=${row.open ?? false}
+					.selected=${row.selected ?? false}
+					.disableSelection=${Boolean(row.disableSelection)}
+					.state=${row.state ?? "default"}
+					@click=${(e: PointerEvent) => this.handleRowClick(row, e)}
+					@toggle-row=${(e: CustomEvent) => this.toggleRowDetails(row, e)}
+					@selected-row=${(e: CustomEvent) => this.handleRowSelection(row, e)}
+				>
+					${getDetailsSlot()}
+					${Object.entries(this.data.header).map(columnHeader => {
+						let width = undefined;
+						let selected = false;
+						let sticky = undefined;
+						let actions = undefined;
+						if (typeof columnHeader[1] === "object") {
+							if (columnHeader[1].width) {
+								width = columnHeader[1].width;
+							}
+							selected = columnHeader[1].selected ?? false;
+							sticky = columnHeader[1].sticky;
+						}
+						const cell = row.data[columnHeader[0]];
 
-					actions = cell.actions;
+						actions = cell.actions;
 
-					return html`<f-tcell
-						.selected=${selected}
-						.width=${width}
-						.actions=${actions}
-						?sticky-left=${ifDefined(sticky)}
-						><f-text inline .highlight=${this.searchTerm}
-							>${this.getCellTemplate(row.data[columnHeader[0]])}</f-text
-						></f-tcell
-					>`;
-				})}
-			</f-trow>`;
-		});
+						let highlightTerm = columnHeader[0] === this.searchScope ? this.searchTerm : null;
+						if (this.searchScope === "all") {
+							highlightTerm = this.searchTerm;
+						}
+
+						return html`<f-tcell
+							part="cell"
+							.selected=${selected}
+							.width=${width}
+							.actions=${actions}
+							?sticky-left=${ifDefined(sticky)}
+							><f-text style="height:100%" inline .highlight=${highlightTerm}
+								>${this.getCellTemplate(row.data[columnHeader[0]])}</f-text
+							></f-tcell
+						>`;
+					})}
+				</f-trow>`;
+			}
+		);
 	}
 
 	get filteredRows() {
@@ -218,7 +248,7 @@ export class FTableSchema extends FRoot {
 	}
 
 	get searchedRows() {
-		if (this.searchTerm) {
+		if (this.searchScope === "all" && this.searchTerm) {
 			return this.data.rows.filter(row => {
 				return (
 					Object.values(row.data).findIndex(v => {
@@ -242,6 +272,23 @@ export class FTableSchema extends FRoot {
 						return true;
 					}) !== -1
 				);
+			});
+		} else if (this.searchScope !== "all" && this.searchTerm) {
+			return this.data.rows.filter(row => {
+				if (this.searchTerm !== null) {
+					const v = row.data[this.searchScope];
+					if (v !== null) {
+						if (typeof v.value === "object" && v.toString) {
+							return v.toString().toLocaleLowerCase().includes(this.searchTerm.toLocaleLowerCase());
+						} else {
+							return v.value
+								.toString()
+								.toLocaleLowerCase()
+								.includes(this.searchTerm.toLocaleLowerCase());
+						}
+					}
+				}
+				return false;
 			});
 		}
 		return this.data?.rows ?? [];
@@ -298,6 +345,7 @@ export class FTableSchema extends FRoot {
 	}
 
 	search(event: CustomEvent) {
+		this.searchScope = event.detail.scope;
 		this.searchTerm = event.detail.value;
 	}
 
@@ -313,6 +361,8 @@ export class FTableSchema extends FRoot {
 					${this.showSearchBar
 						? html`<f-div padding="medium none">
 								<f-search
+									.scope=${["all", ...Object.keys(this.data.header)]}
+									.selected-scope=${this.searchScope}
 									.value=${this.searchTerm}
 									variant="round"
 									@input=${this.search}
@@ -420,6 +470,14 @@ export class FTableSchema extends FRoot {
 		});
 		this.dispatchEvent(rowInputEvent);
 	}
+	handleRowClick(row: FTableSchemaDataRow, _event: PointerEvent) {
+		const rowInputEvent = new CustomEvent("row-click", {
+			detail: row,
+			bubbles: true,
+			composed: true
+		});
+		this.dispatchEvent(rowInputEvent);
+	}
 
 	toggleRowDetails(row: FTableSchemaDataRow, _event: CustomEvent) {
 		row.open = !row.open;
@@ -475,10 +533,10 @@ export class FTableSchema extends FRoot {
 		if (cell && typeof cell === "object" && cell.value && cell.template) {
 			return cell.template();
 		} else if (cell && typeof cell === "object" && cell.value) {
-			return cell.value;
+			return html`<f-text>${cell.value}</f-text>`;
 		}
 
-		return cell;
+		return html`<f-text>${cell}</f-text>`;
 	}
 }