+++ packages/flow-table/src/components/f-tcell/f-tcell.ts
@@ -1,18 +1,22 @@
 import { html, nothing, PropertyValueMap, unsafeCSS } from "lit";
-import { ifDefined } from "lit-html/directives/if-defined.js";
 import { property, query, state } from "lit/decorators.js";
 import { FCheckbox, FDiv, FIconButton, FIcon, FRadio, FRoot, flowElement } from "@cldcvr/flow-core";
 import { FTableSelectable } from "../f-table/f-table";
 import eleStyle from "./f-tcell.scss";
+import { FTrowChevronPosition } from "../f-trow/f-trow";
+import { createRef, ref } from "lit/directives/ref.js";
 
 export type FTcellAction = {
 	icon: string;
-	onClick?: (event: PointerEvent) => void;
-	onMouseOver?: (event: MouseEvent) => void;
-	onMouseLeave?: (event: MouseEvent) => void;
+	id: string;
+	tooltip?: string;
+	onClick?: (event: PointerEvent, element?: FIconButton) => void;
+	onMouseOver?: (event: MouseEvent, element?: FIconButton) => void;
+	onMouseLeave?: (event: MouseEvent, element?: FIconButton) => void;
 };
 
 export type FTcellActions = FTcellAction[];
+
 @flowElement("f-tcell")
 export class FTcell extends FRoot {
 	/**
@@ -43,7 +47,13 @@ export class FTcell extends FRoot {
 	selectable?: FTableSelectable = "none";
 
 	@state()
+	isDisabled = false;
+
+	@state()
 	expandIcon = false;
+
+	@state()
+	expandIconPosition: FTrowChevronPosition = "right";
 
 	@query("f-checkbox")
 	checkbox?: FCheckbox;
@@ -57,15 +67,31 @@ export class FTcell extends FRoot {
 	renderActions() {
 		if (this.actions) {
 			return html`${this.actions.map(ac => {
+				const actionRef = createRef<FIconButton>();
 				return html`<f-icon-button
+					${ref(actionRef)}
+					id=${ac.id}
 					class="f-tcell-actions"
 					size="medium"
 					category="packed"
 					state="neutral"
 					.icon=${ac.icon}
-					@click=${ifDefined(ac.onClick)}
-					@mouseover=${ifDefined(ac.onMouseOver)}
-					@mouseleave=${ifDefined(ac.onMouseLeave)}
+					.tooltip=${ac.tooltip ?? undefined}
+					@click=${(event: PointerEvent) => {
+						if (ac.onClick) {
+							ac.onClick(event, actionRef.value);
+						}
+					}}
+					@mouseover=${(event: MouseEvent) => {
+						if (ac.onMouseOver) {
+							ac.onMouseOver(event, actionRef.value);
+						}
+					}}
+					@mouseleave=${(event: MouseEvent) => {
+						if (ac.onMouseLeave) {
+							ac.onMouseLeave(event, actionRef.value);
+						}
+					}}
 				></f-icon-button>`;
 			})}`;
 		}
@@ -75,14 +101,33 @@ export class FTcell extends FRoot {
 	render() {
 		return html`<f-div aling="middle-left" width="100%" gap="medium"
 			>${this.selectable === "multiple"
-				? html`<f-checkbox @input=${this.handleSelection}></f-checkbox>`
+				? html`<f-checkbox
+						?disabled=${this.isDisabled}
+						@input=${this.handleSelection}
+				  ></f-checkbox>`
 				: nothing}
 			${this.selectable === "single"
-				? html`<f-radio @input=${this.handleSelection} class="cell-radio"></f-radio>`
-				: nothing} <slot></slot>
+				? html`<f-radio
+						?disabled=${this.isDisabled}
+						@input=${this.handleSelection}
+						class="cell-radio"
+				  ></f-radio>`
+				: nothing}
+			${this.expandIcon && this.expandIconPosition === "left"
+				? html`
+						<f-icon-button
+							size="medium"
+							category="packed"
+							class="row-toggler"
+							state="neutral"
+							@click=${this.toggleDetails}
+							icon="i-chevron-down"
+						></f-icon-button>
+				  `
+				: nothing}<slot></slot>
 			<f-div class="details-toggle" width="hug-content" gap="medium" align="top-right">
 				${this.renderActions()}
-				${this.expandIcon
+				${this.expandIcon && this.expandIconPosition === "right"
 					? html`
 							<f-icon-button
 								size="medium"
@@ -98,7 +143,8 @@ export class FTcell extends FRoot {
 		</f-div>`;
 	}
 
-	setSelection(value = false) {
+	setSelection(value = false, isDisabled = false) {
+		this.isDisabled = isDisabled;
 		if (value) {
 			if (this.checkbox) {
 				this.checkbox.value = "checked";