+++ packages/flow-core/CHANGELOG.md
@@ -4,11 +4,44 @@
 
 # Change Log
 
-## [1.17.4] - 2023-07-03
+## [1.18.0] - 2023-07-26
+### Features
 
+- `--color-input-default` color token introduced and used in all inputs.
+
+### Bug Fixes
+- `f-tag` not rendering if `counter="0"`
+
+## [1.17.9] - 2023-07-25
 ### Improvements
 
-- `f-text`: editable feature added.
+- `data-qa-remove-emoji` attribute added in `f-emoji-picker` for automation.
+## [1.17.8] - 2023-07-21
+### Improvements
+
+- `disabled` behavior updated for `f-date-time-picker`, `f-emoji-picker`,`f-icon-button`,`f-search`.
+
+## [1.17.7] - 2023-07-20
+### Bug Fixes
+
+- `f-select` : `disabled` option css fixed.
+- `f-form-builder` not taking full width in `f-div` direction `column`.
+
+## [1.17.6] - 2023-07-11
+### Bug Fixes
+
+- tooltip removal logic updated (when value switch from string to null or undefined)
+
+## [1.17.5] - 2023-07-11
+### Bug Fixes
+
+- `f-search`: `scope` dropdown width updated.
+- disabled css updated for elements in shadowRoot.
+## [1.17.4] - 2023-07-05
+### Bug Fixes
+
+- `f-select`: `create-option` not working in formbuilder.
+- `f-select` : searched option not getting selected in formbuilder.
 
 ## [1.17.3] - 2023-07-03
 ### Improvements